requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cffi==1.17.0.dev0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.0
greenlet==0.4.13
h11==0.14.0
hpy==0.9.0
httptools==0.6.1
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
readline==6.2.4.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
watchfiles==0.23.0
websockets==12.0
